refactor: Move to js modules
