+++ cycle_analytics/templates/adders/event.html
@@ -5,7 +5,6 @@
   integrity="sha256-kLaT2GOSpHechhsozzB+flnD+zUyjE2LlfWPgU04xyI=" crossorigin="" />
 <script src="https://unpkg.com/leaflet@1.9.3/dist/leaflet.js"
   integrity="sha256-WBkoXOwTeyKclOHuWtc+i2uENFpDZ9YPdf5Hf+D7ewM=" crossorigin=""></script>
-<script src="{{ url_for('static', filename='map_utils.js') }}"></script>
 
 {% endblock %}
 
@@ -68,11 +67,7 @@
     The location can be inserted by hand or selected via the map available after clicking the button below.
   </div>
   <div>
-    <button id="btn_open_map" type="button" class="btn btn-success w-100" {% if map_data is none %}
-      onclick="show_map_for_form('map','btn_open_map','400',{{init_map_view[0]}},{{init_map_view[1]}},{{init_map_view[2]}})"
-      {% else %} onclick="show_map_for_form_path('map','btn_open_map','600',
-                                                       '{{ map_data.paths[0].latitudes}}',
-                                                       '{{ map_data.paths[0].longitudes }}')" {% endif %}>
+    <button id="btn_open_map"  class="btn btn-success w-100">
       Open Map
     </button>
   </div>
@@ -83,3 +78,25 @@
 
 
 {% endblock %}
+
+{% block body_scripts %}
+<script type="module" src="{{ url_for('static', filename='map_utils.js') }}"></script>
+<script type="module">
+  import {show_map_for_form_path, show_map_for_form} from "{{ url_for('static', filename='map_utils.js') }}";
+
+  {% if map_data is none %}
+  document.getElementById("btn_open_map").onclick = function() {
+    show_map_for_form('map','btn_open_map','400',{{init_map_view[0]}},{{init_map_view[1]}},{{init_map_view[2]}}, [])
+  }
+  {% else %}
+  document.getElementById("btn_open_map").onclick = function() {
+    show_map_for_form_path('map','btn_open_map','600',
+                           '{{ map_data.paths[0].latitudes}}',
+                           '{{ map_data.paths[0].longitudes }}')
+  }
+  {% endif %}
+</script>
+
+{% endblock %}
+
+
