+++ cycle_analytics/templates/adders/location.html
@@ -5,7 +5,7 @@
     integrity="sha256-kLaT2GOSpHechhsozzB+flnD+zUyjE2LlfWPgU04xyI=" crossorigin="" />
 <script src="https://unpkg.com/leaflet@1.9.3/dist/leaflet.js"
     integrity="sha256-WBkoXOwTeyKclOHuWtc+i2uENFpDZ9YPdf5Hf+D7ewM=" crossorigin=""></script>
-<script src="{{ url_for('static', filename='map_utils.js') }}"></script>
+<script type="module" src="{{ url_for('static', filename='map_utils.js') }}"></script>
 
 {% endblock %}
 
@@ -49,7 +49,8 @@
 {% endblock %}
 
 {% block body_scripts %}
-<script>
+<script type="module">
+    import {EventMarker, show_map_for_form} from "{{ url_for('static', filename='map_utils.js') }}";
     show_map_for_form('map', null, '600', {{ init_map_view[0]}}, {{ init_map_view[1]}}, {{ init_map_view[2]}}, [{% for marker in location_markers %}
     new EventMarker({{ marker.latitude }}, {{ marker.longitude }}, "{{marker.color}}", {{ marker.color_idx }}, "{{marker.popup_text | safe}}")
     {% if loop.index < (location_markers | length) %}, {% endif %}
