+++ cycle_analytics/templates/events.html
@@ -85,8 +85,10 @@
     });
 </script>
 {% if located_events|length > 0 %}
-<script src="{{ url_for('static', filename='map_utils.js') }}"></script>
-<script>
+<script type="module" src="{{ url_for('static', filename='map_utils.js') }}"></script>
+<script type="module">
+    import {PolyLineData,EventMarker,show_map_with_markers} from "{{ url_for('static', filename='map_utils.js') }}";
+
     show_map_with_markers("map",
         [{% for marker in located_events %}
     new EventMarker({{ marker.latitude }}, {{ marker.longitude }}, "{{marker.color}}", {{ marker.color_idx }}, "{{marker.popup_text | safe}}")
